+++ Client/src/Components/DatasetUpload/DatasetValidationSchema.tsx
@@ -1,12 +1,6 @@
 import * as Yup from 'yup'
 
-const integerMessage = (fieldName: string): string => {
-  return `${fieldName} should be an integer`
-}
-
-const requiredMessage = (fieldName: string): string => {
-  return `${fieldName} is a required field`
-}
+import { integerMessage, requiredMessage } from '../../Common/Helpers/ValidationHelpers'
 
 const referenceValidationSchema = Yup.object().shape({
   year: Yup.number().integer(integerMessage('Year')).required(requiredMessage('Year')).test('len', 'Year must be exactly 4 characters', val => val && val.toString().length === 4),
