+++ Client/src/Components/Utils/Hooks/FetchHooks.tsx
@@ -1,0 +1,18 @@
+import { useEffect, useState } from "react"
+
+import { callGetAllDimensions } from "../../../Remote/Endpoints/DimensionsEndpoint"
+
+export const useDimensions = () => {
+  const [dimensions, setDimensions] = useState([])
+
+  useEffect(() => {
+    const callListDimensions = async () => {
+      const dimensions = await callGetAllDimensions()
+      setDimensions(dimensions)
+    }
+
+    callListDimensions()
+  }, [])
+
+  return { dimensions }
+}
