+++ Client/src/Remote/Endpoints/DimensionsEndpoint.tsx
@@ -27,12 +27,11 @@ export const callDeleteDimension = async (dimensionId: number): Promise<any> => 
   }
 }
 
-// Needs to be talked with frontend
 export const callGetAllDimensions = async (): Promise<IDimensionModel[]> => {
-    const result: IDimensionModel[] = await get(dimensionRoute).json()
-    return result
-    // if (result == 'Success') {
-    // Snackbar has been added
-    // SnackbarUtils.success(`New dimension has been added successfully`)
-    // }
+  const result: IDimensionModel[] = await get(dimensionRoute).json()
+  return result
+  // if (result == 'Success') {
+  // Snackbar has been added
+  // SnackbarUtils.success(`New dimension has been added successfully`)
+  // }
 }
